state is lifted
